Update Message.java
Added a function to read the different types of messages. Also, created classes for different types of messages which I will upload next: HaveMessage, PieceMessage, RequestMessage, BitfieldMessage, [choke, unchoke, interested, uninterested]. We still need to have a function which will write a specified message to an outputStream. My bedtime has passed so I can do this in the morning :)